Move point functions to their own file
