+++ src/runtime/length.h
@@ -3,6 +3,7 @@
 
 #include <stdlib.h>
 #include <stdbool.h>
+#include "runtime/point.h"
 #include "tree_sitter/runtime.h"
 
 typedef struct {
@@ -10,46 +11,6 @@ typedef struct {
   size_t chars;
   TSPoint extent;
 } TSLength;
-
-static inline TSPoint ts_point_add(TSPoint a, TSPoint b) {
-  if (b.row > 0)
-    return (TSPoint){a.row + b.row, b.column};
-  else
-    return (TSPoint){a.row, a.column + b.column};
-}
-
-static inline TSPoint ts_point_sub(TSPoint a, TSPoint b) {
-  if (a.row > b.row)
-    return (TSPoint){a.row - b.row, a.column};
-  else
-    return (TSPoint){0, a.column - b.column};
-}
-
-static inline bool ts_point_lte(TSPoint a, TSPoint b) {
-  return (a.row < b.row) || (a.row == b.row && a.column <= b.column);
-}
-
-static inline bool ts_point_lt(TSPoint a, TSPoint b) {
-  return (a.row < b.row) || (a.row == b.row && a.column < b.column);
-}
-
-static inline bool ts_point_eq(TSPoint a, TSPoint b) {
-  return a.row == b.row && a.column == b.column;
-}
-
-static inline TSPoint ts_point_min(TSPoint a, TSPoint b) {
-  if (a.row < b.row || (a.row == b.row && a.column < b.column))
-    return a;
-  else
-    return b;
-}
-
-static inline TSPoint ts_point_max(TSPoint a, TSPoint b) {
-  if (a.row > b.row || (a.row == b.row && a.column > b.column))
-    return a;
-  else
-    return b;
-}
 
 static inline bool ts_length_is_unknown(TSLength self) {
   return self.bytes > 0 && self.chars == 0;
